Window displays... should have particles but nah

+++ ParticleSystem/ParticleSystem/Defines.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
-typedef double p_type;	//type to use for all particle stuff
+typedef float p_type;	//type to use for all particle stuff
 
 #define USE_CPU
